+++ db/schema.rb
@@ -1488,7 +1488,7 @@ ActiveRecord::Schema[7.2].define(version: 2026_07_25_000000) do
     t.string "name", null: false
     t.integer "decimals", default: 18, null: false
     t.string "currency", default: "USD", null: false
-    t.decimal "quantity", precision: 32, scale: 18, default: "0.0", null: false
+    t.decimal "quantity", precision: 36, scale: 18, default: "0.0", null: false
     t.decimal "current_balance", precision: 19, scale: 4, default: "0.0", null: false
     t.jsonb "institution_metadata"
     t.jsonb "raw_payload"
